requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
django==5.2.8
ecdsa==0.19.1
fastapi==0.121.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2024.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.1.2
idna==3.11
jinja2==3.1.6
jiter==0.12.0
json-repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-neo4j==0.6.0
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langchain-tools==0.1.34
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy-extensions==1.1.0
neo4j==5.28.2
neo4j-graphrag==1.10.1
numpy==2.3.4
openai==2.7.2
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
passlib==1.7.4
pillow==12.0.0
propcache==0.4.1
psycopg==3.2.12
psycopg-binary==3.2.12
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pypdf==6.2.0
python-decouple==3.8
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
smolagents==1.22.0
sniffio==1.3.1
sqlalchemy==2.0.44
sqlparse==0.5.3
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typer-slim==0.20.0
types-pyyaml==6.0.12.20250915
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
